requirements.txt
streamlit
requests
pypdf
faiss-cpu
langchain
langchain_community
sentence-transformers
langchain-google-genai
python-dotenv
